Changed time_scale from double to int.  Not sure how it got to be a double in the first place.

+++ src/Oscilloscope.java
@@ -51,7 +51,7 @@ class Oscilloscope extends JFrame implements
 
 	// Number of time steps per scope pixel
 	// Variable named 'speed' in the original Scope class
-	double time_scale;
+	int time_scale;
 	double voltage_range;
 	double current_range;
 	double power_range;
